clean up plotting code and move to a new module
