--- q2_moshpit/assets/checkm/results.tsv
@@ -1,3 +1,0 @@
-sample_id	bin_id	marker_lineage	genomes	markers	marker_sets	count0	count1	count2	count3	count4	count5_or_more	completeness	contamination	gc	gc_std	genome_size	ambiguous_bases	scaffolds	contigs	longest_scaffold	longest_contig	n50_scaffolds	n50_contigs	mean_scaffold_length	mean_contig_length	coding_density	translation_table	predicted_genes	gcn0	gcn1	gcn2	gcn3	gcn4	gcn5_or_more
-sample1	bin1	c__Gammaproteobacteria	1164	275	174	158	117	0	0	0	0	41.371328224776505	0.0	0.3255241194938157	0.018212437862829587	755219	0	198	198	12218	12218	3913	3913	3814.2373737373737	3814.2373737373737	0.9167738099809459	11	894	['PF00572', 'PF01351', 'TIGR00418', 'TIGR00922', 'PF00237', 'PF01765', 'PF02559', 'TIGR00064', 'TIGR02209', 'PF00697', 'PF06574', 'PF03147', 'PF00013', 'PF01386', 'TIGR01499', 'PF03054', 'TIGR01455', 'PF00471', 'TIGR03346', 'TIGR00186', 'PF05188', 'PF10458', 'TIGR00717', 'PF00218', 'PF00035', 'PF00298', 'PF02873', 'PF14450', 'PF02132', 'PF05491', 'PF03461', 'TIGR00460', 'TIGR01059', 'TIGR02800', 'PF01043', 'PF04052', 'TIGR03284', 'PF10396', 'TIGR03263', 'PF12631', 'PF01264', 'PF02978', 'PF00333', 'PF00623', 'TIGR00742', 'PF03484', 'TIGR00878', 'TIGR01063', 'PF01795', 'PF04999', 'TIGR00406', 'PF00334', 'PF01119', 'TIGR00325', 'TIGR01510', 'PF04983', 'PF01281', 'PF00189', 'PF02033', 'PF03948', 'PF02699', 'PF00380', 'TIGR01830', 'PF02843', 'PF12169', 'PF01071', 'PF00162', 'TIGR00436', 'PF02410', 'PF02403', 'TIGR02350', 'PF00958', 'PF00318', 'TIGR00634', 'PF01134', 'PF00562', 'PF01715', 'PF11987', 'PF03799', 'PF00344', 'PF03719', 'PF00627', 'PF01746', 'PF04560', 'TIGR00020', 'TIGR00445', 'TIGR02075', 'PF09312', 'PF01176', 'PF00181', 'PF03726', 'PF07517', 'TIGR00214', 'TIGR00336', 'PF05190', 'TIGR01055', 'TIGR01302', 'PF02491', 'PF01624', 'PF04359', 'PF02606', 'TIGR01740', 'TIGR01135', 'PF03946', 'PF04362', 'PF04977', 'PF02527', 'PF01509', 'PF00830', 'TIGR01394', 'PF08459', 'PF00252', 'PF00276', 'TIGR01062', 'PF04354', 'PF01250', 'PF02844', 'PF01025', 'PF00889', 'PF04386', 'TIGR01144', 'TIGR00422', 'PF00828', 'TIGR00442', 'TIGR02393', 'PF01084', 'PF07497', 'PF01687', 'PF06144', 'PF00327', 'TIGR00416', 'PF01649', 'PF03947', 'TIGR01133', 'TIGR00539', 'PF00203', 'PF04997', 'PF05000', 'PF05698', 'PF05697', 'TIGR00409', 'TIGR01406', 'TIGR01143', 'PF02096', 'TIGR01854', 'PF00584', 'PF13932', 'PF02224', 'PF06421', 'PF02367', 'TIGR00547', 'PF01808', 'TIGR01145', 'PF01245', 'PF00861', 'PF01416', 'PF01192', 'PF02575']	['PF04546', 'TIGR01951', 'PF00410', 'PF06835', 'TIGR03723', 'PF08529', 'PF11734', 'PF00163', 'PF00164', 'PF06418', 'TIGR01083', 'TIGR00755', 'PF02767', 'PF07499', 'PF02130', 'TIGR03594', 'PF07516', 'TIGR00438', 'PF04760', 'TIGR00392', 'TIGR01134', 'PF01409', 'TIGR03533', 'TIGR00668', 'PF02565', 'PF00253', 'PF00338', 'TIGR01173', 'PF00886', 'PF07479', 'TIGR03725', 'TIGR03598', 'TIGR00019', 'PF00573', 'PF00829', 'PF00238', 'PF00366', 'PF03979', 'PF01196', 'PF11967', 'PF03485', 'PF04561', 'TIGR00430', 'PF04998', 'PF02773', 'PF13603', 'TIGR01082', 'PF00687', 'PF08299', 'PF00186', 'TIGR01079', 'PF00684', 'TIGR02614', 'PF00281', 'PF01406', 'PF02075', 'PF00466', 'TIGR02392', 'TIGR01357', 'PF09190', 'TIGR00091', 'PF02684', 'PF02590', 'PF04563', 'TIGR01087', 'TIGR00593', 'PF00411', 'PF02938', 'PF10385', 'PF02772', 'PF01782', 'TIGR01216', 'TIGR02432', 'PF00750', 'PF01783', 'PF13491', 'TIGR00503', 'PF04390', 'PF02092', 'TIGR00595', 'PF03602', 'PF01121', 'PF01018', 'PF00438', 'PF02912', 'TIGR00575', 'PF00542', 'PF02091', 'PF02547', 'TIGR01085', 'PF04127', 'PF02768', 'PF00177', 'PF00712', 'PF00347', 'PF00831', 'PF00453', 'PF13184', 'PF01668', 'TIGR00041', 'TIGR02211', 'PF06071', 'PF01330', 'PF00297', 'TIGR00344', 'TIGR00329', 'PF01195', 'PF01016', 'TIGR01750', 'PF00673', 'PF04565', 'PF03652', 'TIGR01072', 'TIGR00096', 'TIGR03317', 'PF01255', 'PF02576']	[]	[]	[]	[]
-sample2	bin1	k__Bacteria	5449	104	58	89	15	0	0	0	0	19.122257053291534	0.0	0.35717159632687795	0.02014940933075589	280851	0	86	86	5047	5047	3263	3263	3265.7093023255816	3265.7093023255816	0.8821795186771634	11	373	['PF11987', 'PF00623', 'TIGR00459', 'TIGR00084', 'PF04565', 'PF01121', 'PF02367', 'PF00237', 'TIGR00460', 'TIGR02432', 'PF01632', 'TIGR00967', 'PF13184', 'PF08529', 'TIGR03723', 'PF03947', 'PF05697', 'PF00333', 'TIGR00329', 'PF04997', 'PF01195', 'TIGR01079', 'PF00162', 'PF00298', 'PF00164', 'PF04983', 'PF00416', 'PF04563', 'PF00453', 'PF06421', 'PF01250', 'PF00410', 'PF00828', 'TIGR02075', 'PF00347', 'PF00318', 'PF03484', 'PF00312', 'PF05491', 'PF00276', 'PF05000', 'PF00562', 'PF00573', 'PF02912', 'PF01409', 'PF01795', 'PF04560', 'PF00189', 'PF00281', 'TIGR00810', 'PF00673', 'PF01668', 'TIGR00922', 'PF02130', 'PF01509', 'PF04998', 'TIGR00250', 'PF00572', 'PF00687', 'PF01649', 'PF00181', 'PF03946', 'PF00886', 'TIGR03594', 'PF01000', 'PF00252', 'PF00411', 'PF00380', 'PF00366', 'PF01765', 'PF01193', 'PF10385', 'PF01018', 'PF00889', 'PF00177', 'PF04561', 'PF00861', 'PF00238', 'PF01196', 'TIGR00392', 'PF02978', 'PF01245', 'TIGR00755', 'PF06071', 'TIGR00615', 'PF02033', 'PF03719', 'PF00203', 'PF00831']	['PF00297', 'PF12344', 'PF00338', 'PF13603', 'TIGR00019', 'PF01281', 'TIGR00855', 'PF08459', 'PF01016', 'TIGR03263', 'TIGR00344', 'PF00466', 'PF03948', 'PF01746', 'PF00829']	[]	[]	[]	[]
