+++ q2_moshpit/checkm/checkm.py
@@ -13,12 +13,12 @@ from copy import deepcopy
 from distutils.dir_util import copy_tree
 from zipfile import ZipFile
 
-import altair as alt
 import pandas as pd
 import pkg_resources
 import q2templates
 
 from q2_moshpit._utils import _process_common_input_params, run_command
+from q2_moshpit.checkm.plots import _draw_final_plot
 from q2_moshpit.checkm.utils import _process_checkm_arg, _get_plots_per_sample
 
 from q2_types_genomics.per_sample_data._format import \
@@ -142,249 +142,6 @@ def _parse_single_checkm_report(
     return df
 
 
-def _draw_plots(df: pd.DataFrame) -> dict:
-    # rename columns for better plot labels
-    col_names = {
-        'count0': '0', 'count1': '1', 'count2': '2',
-        'count3': '3', 'count4': '4', 'count5_or_more': '5+'
-    }
-    df = df.rename(columns=col_names, inplace=False)
-
-    # convert genome size to Mbp
-    df['genome_size'] = df['genome_size'] / 10**6
-
-    # prepare required selectors
-    sample_ids = df['sample_id'].unique()
-    sample_dropdown = alt.binding_select(
-        options=sample_ids,
-        name='Sample ID  '
-    )
-    sample_selection = alt.selection_single(
-        fields=['sample_id'],
-        bind=sample_dropdown,
-        init={'sample_id': sample_ids[0]}
-    )
-
-    opacity_selection = alt.selection_single(
-        fields=['opacity'],
-        bind=alt.binding_range(
-            min=0.1, max=1, name='Opacity: other samples  '
-        ),
-        init={'opacity': 0.15}
-    )
-
-    base = alt.Chart(df).transform_calculate(
-        opacity=opacity_selection.opacity
-    ).transform_fold(
-        list(col_names.values())
-    )
-
-    # prep and concatenate all plots
-    final_plot = _concatenate_plots(
-        completeness_plot=_prep_completeness_plot(
-            base, opacity_selection, sample_selection
-        ),
-        gc_plot=_prep_gc_plot(base, opacity_selection, sample_selection),
-        marker_plot=_prep_markers_plot(base, sample_selection),
-        contig_plots=_prep_contig_plots(base, sample_selection),
-        genes_plot=_prep_genes_plot(base, opacity_selection, sample_selection),
-        contig_count_plot=_prep_contig_count_plot(
-            base, opacity_selection, sample_selection
-        )
-    )
-
-    return final_plot.to_dict()
-
-
-def _concatenate_plots(
-        completeness_plot, gc_plot, marker_plot, contig_plots,
-        genes_plot, contig_count_plot
-):
-    plot = alt.vconcat(
-        alt.hconcat(
-            completeness_plot, gc_plot, spacing=40
-        ),
-        alt.hconcat(
-            genes_plot, contig_count_plot, spacing=40
-        ),
-        marker_plot,
-        *contig_plots.values(),
-        spacing=40
-    ).resolve_scale(
-        color='independent'
-    ).configure_axis(
-        labelFontSize=12, titleFontSize=15
-    ).configure_legend(
-        labelFontSize=12, titleFontSize=14
-    )
-    return plot
-
-
-def _prep_contig_plots(base_plot, sample_selection):
-    contig_plots = {}
-    contig_cols = {
-        'longest_contig': 'Longest contig length [bp]',
-        'n50_contigs': 'N50 contigs [bp]',
-        'mean_contig_length': 'Mean contig length [bp]',
-        'ambiguous_bases': 'Count of ambiguous bases'
-    }
-    for y, title in contig_cols.items():
-        contig_plots[y] = \
-            base_plot.mark_bar().encode(
-                x=alt.X('bin_id:N', title='Bin'),
-                y=alt.Y(f'{y}:Q', title=title),
-            ).add_selection(
-                sample_selection
-            ).transform_filter(
-                sample_selection
-            ).properties(
-                width=880, height=250
-            )
-    return contig_plots
-
-
-def _prep_gc_plot(base_plot, opacity_selection, sample_selection):
-    plot_gc = base_plot.mark_point(
-        size=80, filled=True, opacity=0.7,
-        strokeWidth=0.2, strokeOpacity=0.8, stroke='lightgrey'
-    ).encode(
-        x=alt.X('gc:Q', title='GC content'),
-        y=alt.Y('coding_density:Q', title='Coding density'),
-        color=alt.Color(
-            'sample_id:N',
-            scale=alt.Scale(scheme='tableau10'),
-            legend=alt.Legend(title='Sample ID')
-        ),
-        tooltip=[
-            alt.Tooltip('sample_id:N', title='Sample ID'),
-            alt.Tooltip('bin_id:N', title='Bin ID'),
-            alt.Tooltip('coding_density:Q', title='Coding density'),
-            alt.Tooltip('gc:Q', title='GC content'),
-        ],
-        opacity=alt.condition(
-            sample_selection, alt.value(1),
-            alt.Opacity('opacity:Q', scale=None)
-        )
-    ).add_selection(
-        sample_selection, opacity_selection
-    ).properties(
-        width=400, height=400
-    ).interactive()
-    return plot_gc
-
-
-def _prep_genes_plot(base_plot, opacity_selection, sample_selection):
-    plot_genes = base_plot.mark_point(
-        size=80, filled=True, opacity=0.7,
-        strokeWidth=0.2, strokeOpacity=0.8, stroke='lightgrey'
-    ).encode(
-        x=alt.X('genome_size:Q', title='Genome size [Mbp]'),
-        y=alt.Y('predicted_genes:Q', title='Count of predicted genes'),
-        color=alt.Color(
-            'sample_id:N',
-            scale=alt.Scale(scheme='tableau10'),
-            legend=alt.Legend(title='Sample ID')
-        ),
-        tooltip=[
-            alt.Tooltip('sample_id:N', title='Sample ID'),
-            alt.Tooltip('bin_id:N', title='Bin ID'),
-            alt.Tooltip('genome_size:Q', title='Genome size [Mbp]'),
-            alt.Tooltip('predicted_genes:Q', title='Predicted genes'),
-        ],
-        opacity=alt.condition(
-            sample_selection, alt.value(1),
-            alt.Opacity('opacity:Q', scale=None)
-        )
-    ).add_selection(
-        sample_selection, opacity_selection
-    ).properties(
-        width=400, height=400
-    ).interactive()
-    return plot_genes
-
-
-def _prep_contig_count_plot(base_plot, opacity_selection, sample_selection):
-    plot_contigs = base_plot.mark_point(
-        size=80, filled=True, opacity=0.7,
-        strokeWidth=0.2, strokeOpacity=0.8, stroke='lightgrey'
-    ).encode(
-        x=alt.X('contigs:Q', title='Count of contigs'),
-        y=alt.Y('genome_size:Q', title='Genome size [Mbp]'),
-        color=alt.Color(
-            'sample_id:N',
-            scale=alt.Scale(scheme='tableau10'),
-            legend=alt.Legend(title='Sample ID')
-        ),
-        tooltip=[
-            alt.Tooltip('sample_id:N', title='Sample ID'),
-            alt.Tooltip('bin_id:N', title='Bin ID'),
-            alt.Tooltip('genome_size:Q', title='Genome size [Mbp]'),
-            alt.Tooltip('contigs:Q', title='Contigs'),
-        ],
-        opacity=alt.condition(
-            sample_selection, alt.value(1),
-            alt.Opacity('opacity:Q', scale=None)
-        )
-    ).add_selection(
-        sample_selection, opacity_selection
-    ).properties(
-        width=400, height=400
-    ).interactive()
-    return plot_contigs
-
-
-def _prep_markers_plot(base_plot, sample_selection):
-    plot_markers = base_plot.mark_bar().encode(
-        x=alt.X('bin_id', title='Bin'),
-        y=alt.Y('sum(value):Q', title='Marker count'),
-        color=alt.Color(
-            'key:N',
-            title='Markers',
-            scale=alt.Scale(scheme='blues'),
-            sort='ascending'
-        )
-    ).add_selection(
-        sample_selection
-    ).transform_filter(
-        sample_selection
-    ).properties(
-        width=880, height=250
-    )
-    return plot_markers
-
-
-def _prep_completeness_plot(base_plot, opacity_selection, sample_selection):
-    plot_completeness = base_plot.mark_point(
-        size=80, filled=True, opacity=0.7,
-        strokeWidth=0.2, strokeOpacity=0.8, stroke='lightgrey'
-    ).encode(
-        x=alt.X('completeness:Q', title='Completeness [%]'),
-        y=alt.Y('contamination:Q', title='Contamination [%]'),
-        color=alt.Color(
-            'sample_id:N',
-            scale=alt.Scale(scheme='viridis'),
-            legend=alt.Legend(title='Sample ID')
-        ),
-        tooltip=[
-            alt.Tooltip('sample_id:N', title='Sample ID'),
-            alt.Tooltip('bin_id:N', title='Bin ID'),
-            alt.Tooltip('marker_lineage:N', title='Marker lineage'),
-            alt.Tooltip('completeness:Q', title='% completeness'),
-            alt.Tooltip('contamination:Q', title='% contamination')
-        ],
-        opacity=alt.condition(
-            predicate=sample_selection,
-            if_true=alt.value(1),
-            if_false=alt.Opacity('opacity:Q', scale=None)
-        )
-    ).add_selection(
-        sample_selection, opacity_selection
-    ).properties(
-        width=400, height=400
-    ).interactive()
-    return plot_completeness
-
-
 def evaluate_bins(
     output_dir: str, bins: MultiMAGSequencesDirFmt,
     db_path: str, reduced_tree: bool = None, unique: int = None,
@@ -442,7 +199,7 @@ def evaluate_bins(
         context = {
             'samples': json.dumps(list(reports.keys())),
             'vega_plots': json.dumps(
-                _draw_plots(checkm_results)
+                _draw_final_plot(checkm_results)
             ),
             **{k: json.dumps(v) for k, v in all_plots.items()}
         }
