+++ q2_moshpit/checkm/tests/data/bin_stats_ext1.tsv
@@ -1,0 +1,2 @@
+bin1	{'marker lineage': 'g__Mycobacterium', '# genomes': 100, '# markers': 693, '# marker sets': 300, '0': 0, '1': 693, '2': 0, '3': 0, '4': 0, '5+': 0, 'Completeness': 100.0, 'Contamination': 0.0, 'GC': 0.6413333033893058, 'GC std': 0.005891840780810075, 'Genome size': 5120665, '# ambiguous bases': 0, '# scaffolds': 14, '# contigs': 14, 'Longest scaffold': 1063123, 'Longest contig': 1063123, 'N50 (scaffolds)': 526051, 'N50 (contigs)': 526051, 'Mean scaffold length': 365761.78571428574, 'Mean contig length': 365761.78571428574, 'Coding density': 0.9221700697077431, 'Translation table': 11, '# predicted genes': 5099, 'GCN0': [], 'GCN1': ['PF09992', 'F03668'], 'GCN2': [], 'GCN3': [], 'GCN4': [], 'GCN5+': []}
+bin2	{'marker lineage': 'o__Pseudomonadales', '# genomes': 185, '# markers': 813, '# marker sets': 308, '0': 97, '1': 707, '2': 9, '3': 0, '4': 0, '5+': 0, 'Completeness': 93.86937557392102, 'Contamination': 1.3095238095238095, 'GC': 0.6328437335198756, 'GC std': 0.021845669875020592, 'Genome size': 5895359, '# ambiguous bases': 0, '# scaffolds': 500, '# contigs': 500, 'Longest scaffold': 59893, 'Longest contig': 59893, 'N50 (scaffolds)': 15797, 'N50 (contigs)': 15797, 'Mean scaffold length': 11790.718, 'Mean contig length': 11790.718, 'Coding density': 0.8790287410826041, 'Translation table': 11, '# predicted genes': 5574, 'GCN0': ['PF00181'], 'GCN1': ['PF04379', 'TIGR01510'], 'GCN2': ['PF09831'], 'GCN3': [], 'GCN4': [], 'GCN5+': []}
