+++ q2_moshpit/checkm/tests/data/bin_stats_ext2.tsv
@@ -1,0 +1,2 @@
+bin1	{'marker lineage': 'c__Alphaproteobacteria', '# genomes': 564, '# markers': 337, '# marker sets': 221, '0': 0, '1': 337, '2': 0, '3': 0, '4': 0, '5+': 0, 'Completeness': 100.0, 'Contamination': 0.0, 'GC': 0.4627558694656988, 'GC std': 0.011898590215179758, 'Genome size': 2033703, '# ambiguous bases': 0, '# scaffolds': 13, '# contigs': 13, 'Longest scaffold': 1011287, 'Longest contig': 1011287, 'N50 (scaffolds)': 164033, 'N50 (contigs)': 164033, 'Mean scaffold length': 156438.6923076923, 'Mean contig length': 156438.6923076923, 'Coding density': 0.8715200793822894, 'Translation table': 11, '# predicted genes': 1752, 'GCN0': [], 'GCN1': ['TIGR01079', 'PF02934', 'PF01624'], 'GCN2': [], 'GCN3': [], 'GCN4': [], 'GCN5+': []}
+bin2	{'marker lineage': 'f__Enterobacteriaceae', '# genomes': 157, '# markers': 1005, '# marker sets': 324, '0': 9, '1': 995, '2': 1, '3': 0, '4': 0, '5+': 0, 'Completeness': 99.38271604938272, 'Contamination': 0.07716049382716049, 'GC': 0.5775877063191994, 'GC std': 0.014852537713170508, 'Genome size': 5082416, '# ambiguous bases': 0, '# scaffolds': 39, '# contigs': 39, 'Longest scaffold': 302260, 'Longest contig': 302260, 'N50 (scaffolds)': 177543, 'N50 (contigs)': 177543, 'Mean scaffold length': 130318.35897435897, 'Mean contig length': 130318.35897435897, 'Coding density': 0.8868593204491722, 'Translation table': 11, '# predicted genes': 4762, 'GCN0': ['TIGR00007', 'PF01502'], 'GCN1': ['TIGR02063'], 'GCN2': ['PF13145'], 'GCN3': [], 'GCN4': [], 'GCN5+': []}
